Dialogue
Changes to scene 
1. added a canvas with the dialogue UI (subject to appearance change)
2. added two temporary NPC's to interact with
3. added box collider to the main character, to be picked up by the NPC when you walk near it (so you can toggle chat on and off)

Script changes:
1. Minor changes made to the use-case in PlayerController script, so that when we are in dialogue we cannot move (animations still bugged though, like you'll be stuck in the walking animation if you are walking and start the dialogue. You cannot move anywhere, but you'll be walking in place).

Added new scripts:
Dialogue: This will handle opening the UI and displaying everything in it. 
Dialogue Interaction: This will be unique to each NPC, you can specify custom lines of dialogue. The lines are stored here, and are handled by the Dialogue script. Also handles the onTrigger interactions to detect the player.