Room Pictures
Added a collider to the room numbers, if you press E on them a picture of the room will come up, if you press E again the picture will go away. The character isn't locked in place when you do this though, so don't move. I can add in the lock, but I wasn't sure if we were going change out the character or not.  So I'll just wait.